Failed attempt to build pjmedia with DirectSound on Linux
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@171 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip/src/pjsua/pjsua_core.c
@@ -561,9 +561,9 @@ pj_status_t pjsua_start(void)
 	    /* With the user part. */
 	    len = pj_snprintf(contact, sizeof(contact),
 			      "<sip:%.*s@%.*s:%d>",
-			      sip_uri->user.slen,
+			      (int)sip_uri->user.slen,
 			      sip_uri->user.ptr,
-			      udp_transport->local_name.host.slen,
+			      (int)udp_transport->local_name.host.slen,
 			      udp_transport->local_name.host.ptr,
 			      udp_transport->local_name.port);
 	} else {
@@ -572,7 +572,7 @@ pj_status_t pjsua_start(void)
 
 	    len = pj_snprintf(contact, sizeof(contact),
 			      "<sip:%.*s:%d>",
-			      udp_transport->local_name.host.slen,
+			      (int)udp_transport->local_name.host.slen,
 			      udp_transport->local_name.host.ptr,
 			      udp_transport->local_name.port);
 	}
